chore: move new pipeline collector config generation logic to its own file
